requirements.txt
python-telegram-bot==13.15
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
selenium==4.10.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
webdrivermanager==0.10.0
gunicorn==21.2.0
